User roles: Add entity permissions to test fixtures (#5313)
* Prevented deleting of entity permissions if they are specified in the fixture

* Added entity permissions to test_publishers.yaml

* Added entity permissions to test_api_breakdowns.yaml

* Added write permissions to yaml

* Added entity permissions to test_upload.yaml

* Added entity permissions to test_api.yaml

* Added write permissions to yaml

* Added entity permissions to test_audiences.yaml

* Added entity permissions to test_views.yaml

* Removed unused users arrays from test_e2e.yaml

* Added entity permissions to test_models.yaml

* Removed reference to account.users from test

* Removed users from test_projections.yaml

* Removed users from test_augmenter.yaml

* Removed users from test_daily_statements.yaml

* Added entity permissions to test_navigation.yaml

+++ server/zemauth/features/entity_permission/service.py
@@ -15,6 +15,9 @@ EXTERNAL_REST_API_GROUP_ID = 33
 
 @transaction.atomic
 def refresh_entity_permissions_for_user(user):
+    if len(user.entitypermission_set.all()) > 0:
+        return
+
     user.entitypermission_set.all().delete()
     if _handle_internal_user(user):
         return
